Fixing build and making capture fabric

+++ core/geometry/convexQuickHull.cpp
@@ -182,7 +182,7 @@ bool ConvexQuickHull::faceIsVisible(const Vector3dd &eyePoint, const ConvexQuick
 
 void ConvexQuickHull::addPointsToFaces(HullFace *faces, unsigned long faces_count, const vertices &listVertices, double eps) {
     for (auto vertex : listVertices)
-        for (int i = 0; i < faces_count; i++) {
+        for (unsigned long i = 0; i < faces_count; i++) {
             HullFace face = faces[i];
             if ((vertex != face.plane.p1()) && (vertex != face.plane.p2()) && (vertex != face.plane.p3()) && faceIsVisible(vertex, face, eps)) {
                 faces[i].points.push_back(vertex);
